src/ingestion/sports_scraper_engine.py
"""
Sports Scraper Engine — Scrape structured sports data from public sources.
Sources: Wikipedia, public sports databases.
Respects robots.txt and rate limits.
"""
import logging
import re
from typing import List, Dict, Any, Optional
from datetime import datetime
import httpx
from bs4 import BeautifulSoup

from .base_adapter import BaseSportsAdapter
from .sports_schema import Match, SportType, StandingsEntry, Team

logger = logging.getLogger(__name__)


class SportsScraper(BaseSportsAdapter):
    """
    Scraper for public sports data sources.
    Primarily targets Wikipedia structured tables.
    """

    def __init__(self):
        super().__init__(
            source_name="web-scraper",
            sport=SportType.FOOTBALL,
            cache_ttl=3600,  # 1 hour cache for scraped data
            rate_limit=5,
            rate_window=60,
        )

    async def initialize(self) -> bool:
        self._initialized = True
        logger.info("Sports scraper initialized")
        return True

    async def _fetch_page(self, url: str) -> Optional[str]:
        """Fetch HTML page with rate limiting"""
        await self._check_rate_limit()
        try:
            async with httpx.AsyncClient(
                timeout=15.0,
                headers={"User-Agent": "ATTI-Sports-Bot/1.0 (research)"},
            ) as client:
                resp = await client.get(url)
                if resp.status_code == 200:
                    return resp.text
                return None
        except Exception as e:
            logger.error(f"Scrape failed for {url}: {e}")
            return None

    async def scrape_brasileirao_standings(self, year: int = 2026) -> List[Dict[str, Any]]:
        """Scrape Brasileirão standings from Wikipedia"""
        cache_key = self._cache_key("brasileirao_standings", year=year)
        cached = self._get_cached(cache_key)
        if cached is not None:
            return cached

        url = f"https://pt.wikipedia.org/wiki/Campeonato_Brasileiro_de_Futebol_de_{year}_-_S%C3%A9rie_A"
        html = await self._fetch_page(url)
        if not html:
            return []

        soup = BeautifulSoup(html, "html.parser")
        standings = []

        # Find classification table
        tables = soup.find_all("table", class_="wikitable")
        for table in tables:
            headers = [th.get_text(strip=True) for th in table.find_all("th")]
            if any(h in headers for h in ["Pos", "P", "Pts", "J"]):
                rows = table.find_all("tr")[1:]
                for row in rows:
                    cells = row.find_all(["td", "th"])
                    if len(cells) >= 8:
                        try:
                            entry = {
                                "position": int(cells[0].get_text(strip=True).replace(".", "")),
                                "team": cells[1].get_text(strip=True),
                                "played": int(cells[2].get_text(strip=True) or 0),
                                "won": int(cells[3].get_text(strip=True) or 0),
                                "drawn": int(cells[4].get_text(strip=True) or 0),
                                "lost": int(cells[5].get_text(strip=True) or 0),
                                "goals_for": int(cells[6].get_text(strip=True) or 0),
                                "goals_against": int(cells[7].get_text(strip=True) or 0),
                                "points": int(cells[-1].get_text(strip=True) or 0),
                            }
                            standings.append(entry)
                        except (ValueError, IndexError):
                            continue
                break

        self._set_cached(cache_key, standings)
        return standings

    async def scrape_team_info(self, team_name: str) -> Dict[str, Any]:
        """Scrape team information from Wikipedia"""
        cache_key = self._cache_key("team_info", team=team_name)
        cached = self._get_cached(cache_key)
        if cached is not None:
            return cached

        # Search Wikipedia
        search_url = f"https://pt.wikipedia.org/w/api.php?action=query&list=search&srsearch={team_name} futebol&format=json"
        try:
            async with httpx.AsyncClient(timeout=10.0) as client:
                resp = await client.get(search_url)
                data = resp.json()
                results = data.get("query", {}).get("search", [])
                if not results:
                    return {}

                page_title = results[0]["title"]
                page_url = f"https://pt.wikipedia.org/wiki/{page_title.replace(' ', '_')}"
        except Exception:
            return {}

        html = await self._fetch_page(page_url)
        if not html:
            return {}

        soup = BeautifulSoup(html, "html.parser")
        info = {"name": team_name, "source": page_url}

        # Extract infobox data
        infobox = soup.find("table", class_="infobox")
        if infobox:
            for row in infobox.find_all("tr"):
                header = row.find("th")
                value = row.find("td")
                if header and value:
                    key = header.get_text(strip=True).lower()
                    val = value.get_text(strip=True)
                    if "fundação" in key or "fundado" in key:
                        info["founded"] = val
                    elif "estádio" in key:
                        info["stadium"] = val
                    elif "capacidade" in key:
                        info["capacity"] = val
                    elif "treinador" in key or "técnico" in key:
                        info["coach"] = val

        # Extract first paragraph
        content = soup.find("div", class_="mw-parser-output")
        if content:
            first_p = content.find("p", class_=None)
            if first_p:
                info["summary"] = first_p.get_text(strip=True)[:500]

        self._set_cached(cache_key, info)
        return info

    async def fetch_live_matches(self) -> List[Match]:
        return []

    async def fetch_today_matches(self) -> List[Match]:
        return []

    async def fetch_match_events(self, match_id: str) -> List[Dict[str, Any]]:
        return []

    async def fetch_standings(self, competition_id: str) -> List[Dict[str, Any]]:
        return await self.scrape_brasileirao_standings()
